requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
pandas
python-jose[cryptography]
passlib[bcrypt]
python-multipart
lxml
python-dotenv